Count a compound bar in dotted beats
A step is now a fixed note value in every meter. `Grid` divides the
*quarter*, so four steps is a sixteenth whether the bar is 4/4 or 6/8; it
used to divide the note the denominator names, which made a "sixteenth"
in 6/8 a thirty-second and left the grid twice as fine as everything
placing notes on it believed. A 6/8 bar is twelve steps now, not
twenty-four, and they are the twelve sixteenths the bar actually has.

`steps_per_beat` is derived from the *felt* beat rather than stored, so a
6/8 beat is six steps and a 6/8 bar has two of them. That one number is
what every part asks when it wants to know whether it is on a beat, and
three things fall out of getting it right:

`weight` no longer offers the halves family inside a compound beat. Three
sixteenths into a dotted quarter is a dotted eighth — a syncopation
against the meter, not a position the meter offers — and it was being
handed a real division's weight, which is exactly backwards for the step
in 6/8 that most needs to be heard as a departure.

`swing` goes quiet in compound time. A 6/8 beat already *is* three
eighths, which is where swinging a 4/4 beat is trying to get to; there is
no straight pair left to delay. `is_triplet` had to stop reading
`steps_per_beat` to say so, or it would have called every 6/8 a triplet
grid and switched swing off for the wrong reason.

And a groove is laid over the bar with both beats known: the pattern's
own four steps and the bar's six. In compound time the groove's beat is
stretched over the longer one and each of its steps strikes once, so a
16-step 4/4 pattern in 6/8 comes out as a downbeat and a turnaround —
what two beats of it reduce to, and what a drummer would play.

Every 4/4 number is arithmetically identical, which is why all four
pinned fixtures held.

+++ crates/auris-compose/src/parts/bass.rs
@@ -114,7 +114,7 @@ pub(super) fn bass(
                     .filter(|offset| {
                         kick_at(event.start + grid.tick_of(*offset))
                             || ((first + offset) % per_bar)
-                                .is_multiple_of((grid.steps_per_beat as usize).max(1))
+                                .is_multiple_of(grid.steps_per_beat().max(1))
                     })
                     .collect(),
             },
